core/requirements.txt
fastapi
uvicorn
firebase-admin
requests
python-dotenv
pydantic
google-generativeai
